pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "coreapi>=2.3.3",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework-simplejwt>=5.5.0",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
